mysite/requirements.txt
# Dev environment
pip
autopep8

# App
asgiref==3.7.2
Django==5.0.13
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework==3.15.1
